game.c and game.h were changed for compilations purpose

+++ game.c
@@ -77,7 +77,7 @@ game* game_create() {
 	for (int i = 0; i < 16; i++) {
 		printf("%d\n", i);
 		printf("%d\n", newgame->whites[i].alive);
-		newgame->whites[i].alive = 1;
+		newgame->whites[i].alive = (int)1;
 		printf("A");
 		newgame->whites[i].color = 1;
 		printf("B");
